fix test target clippy lints across reifydb crates

+++ crates/codec/tests/frame/delta/int4.rs
@@ -11,5 +11,5 @@ fn make(v: Vec<i32>) -> FrameColumnData {
 crate::delta_tests! {
 	ascending: (0..200i32).collect::<Vec<_>>(),
 	descending: (0..200i32).rev().collect::<Vec<_>>(),
-	unsorted: (0..200).map(|i| ((i * 7 + 13) % 97) as i32).collect::<Vec<_>>(),
+	unsorted: (0..200).map(|i| (i * 7 + 13) % 97).collect::<Vec<_>>(),
 }
